Structure list outputs

Structure list outputs

Usage

al_structure_areas(areas_list)

al_structure_work(works_list)

al_structure_works(works_list)

Arguments

areas_list

A countries list output produced from a request for countries JSON to the Laws.Africa API.

works_list

A works list outputs produced from a request for works JSON to the Laws.Africa API.

Value

A named list of two tibbles the first of which is for countries information and the second for localities information.